Use fhir-spec-tools
