+++ package.json
@@ -6,10 +6,7 @@
     "check": "npm run test && npm run lint && npm run prettier",
     "dev": "next dev",
     "build": "next build",
-    "build:dates": "ts-node scripts/parsePrimaryDatePath.ts && prettier --write util/primaryDatePaths.ts --loglevel silent",
-    "build:codes": "ts-node scripts/parseCodePath.ts && prettier --write util/codePaths.ts --loglevel silent",
     "build:names": "ts-node scripts/generatePatientNames.ts && prettier --write data/names.ts --loglevel silent",
-    "build:lookups": "npm run build:codes && npm run build:names && npm run build:dates",
     "start": "npx serve@latest out",
     "lint": "next lint && tsc --noEmit",
     "lint:fix": "next lint --fix",
