+++ util/fhir/patient.ts
@@ -1,4 +1,4 @@
-import { ReferencesMap } from '../referencesMap';
+import { patientAttributePaths } from 'fhir-spec-tools/build/data/patient-attribute-paths';
 
 export function getPatientInfoString(patient: fhir4.Patient) {
   return `${getPatientNameString(patient)} (${getPatientDOBString(patient)})`;
@@ -14,12 +14,12 @@ export function getPatientNameString(patient: fhir4.Patient) {
 
 export function getResourcePatientReference(resource: any, dr: fhir4.DataRequirement, patientId: string | null) {
   // determine if we should add a reference to the patient
-  if (ReferencesMap[dr.type] && patientId) {
+  if (patientAttributePaths[dr.type] && patientId) {
     // add if subject is in the list otherwise add it on the first one
-    if (ReferencesMap[dr.type].includes('subject')) {
+    if (patientAttributePaths[dr.type].includes('subject')) {
       resource.subject = { reference: `Patient/${patientId}` };
-    } else if (ReferencesMap[dr.type][0] && !ReferencesMap[dr.type][0].includes('.')) {
-      resource[ReferencesMap[dr.type][0]] = { reference: `Patient/${patientId}` };
+    } else if (patientAttributePaths[dr.type][0] && !patientAttributePaths[dr.type][0].includes('.')) {
+      resource[patientAttributePaths[dr.type][0]] = { reference: `Patient/${patientId}` };
     }
   }
 }
